Restrict vstacking to OrderedSink only.
