+++ crates/polars-pipe/src/operators/chunks.rs
@@ -28,24 +28,7 @@ impl DataChunk {
 }
 
 pub(crate) fn chunks_to_df_unchecked(chunks: Vec<DataChunk>) -> DataFrame {
-    let mut combiner = SemicontiguousVstacker::default();
-    let mut frames_iterator = chunks
-        .into_iter()
-        .flat_map(|c| combiner.add(c.data))
-        .peekable();
-    if frames_iterator.peek().is_some() {
-        let mut result = accumulate_dataframes_vertical_unchecked(frames_iterator);
-        if let Some(df) = combiner.finish() {
-            let _ = result.vstack_mut(&df);
-        }
-        result
-    } else {
-        // The presumption is that this function is never called with empty
-        // data, cause that'll cause accumulate_dataframes_vertical_unchecked to
-        // error, so if we haven't gotten any data we can safely assume it's in
-        // the combiner buffer.
-        combiner.finish().unwrap()
-    }
+    accumulate_dataframes_vertical_unchecked(chunks.into_iter().map(|c| c.data))
 }
 
 /// Combine `DataFrame`s into sufficiently large contiguous memory allocations
